% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_column_S01.R
\name{append_column_S01}
\alias{append_column_S01}
\title{Append Column S01 (Institution)}
\usage{
append_column_S01(input_df = fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element S01 appended.
}
\description{
Append Column S01 (Institution)
}
\examples{
append_column_S01()

}
