% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_column_S02.R
\name{append_column_S02}
\alias{append_column_S02}
\title{Append Column S02 (Year, Term, & Extract)}
\usage{
append_column_S02(input_df = fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element S02 appended.
}
\description{
Append Column S02 (Year, Term, & Extract)
}
\examples{
append_column_S02()

}
