% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_column_S12.R
\name{append_column_S12}
\alias{append_column_S12}
\title{Append Column S12 (Birth Date)}
\usage{
append_column_S12(input_df = fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element S14 appended.
}
\description{
Append Column S12 (Birth Date)
}
\examples{
append_column_S12()

}
