% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_column_S14.R
\name{append_column_S14}
\alias{append_column_S14}
\title{Append Column S14 (Ethnic Origin)}
\usage{
append_column_S14(input_df = fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element S14 appended.
}
\description{
Append Column S14 (Ethnic Origin)
}
\examples{
append_column_S14()

}
