% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_02}
\alias{b_02}
\title{Calculate USHE Element b_02 (Building Location Code)}
\usage{
b_02(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_location_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_02 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_02 (Building Location Code)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Location Code
\item FIELD NAME: b_location
\item FIELD FORMAT: Varchar, 3 Characters
\item DEFINITION: This field has been established to identify the campus location or other site  where the space is located.
}
}
\examples{
b_02()

}
