% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_03}
\alias{b_03}
\title{Calculate USHE Element b_03 (Building Ownership Code)}
\usage{
b_03(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_ownership_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_03 (Building Ownership Code)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Ownership Status
\item FIELD NAME: b_ownership
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: This field has been established to identify the type of ownership of space included in the inventory including space that is not owned by the institution but leased for its use from other parties.
}
}
\examples{
b_03()

}
