% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_04}
\alias{b_04}
\title{Calculate USHE Element b_04 (Year Submission)}
\usage{
b_04(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (submission_year).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_04 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_04 (Year Submission)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Submission Year
\item FIELD NAME: b_year
\item FIELD FORMAT: Varchar, 4 Characters (YYYY format)
\item DEFINITION:The fiscal year in which the Space Inventory - Building inventory was conducted.
}
}
\examples{
b_04()

}
