% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_05}
\alias{b_05}
\title{Calculate USHE Element b_05 (Building Name)}
\usage{
b_05(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_name).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_05 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_05 (Building Name)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Name
\item FIELD NAME: b_name
\item FIELD FORMAT: Varchar, 255 Characters
\item DEFINITION: A unique identifier assigned by the institution to the specific building.
}
}
\examples{
b_05()

}
