% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_06}
\alias{b_06}
\title{Calculate USHE Element b_06 (Building Name)}
\usage{
b_06(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_number).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_06 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_06 (Building Name)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Number
\item FIELD NAME: b_number
\item FIELD FORMAT: Varchar, 7 Characters
\item DEFINITION: A unique identifier assigned by the institution for the building where the room is located.
}
}
\examples{
b_06()

}
