% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_07}
\alias{b_07}
\title{Calculate USHE Element b_07 (Building Abbreviation)}
\usage{
b_07(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_abbrv).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_07 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_07 (Building Abbreviation)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Abbreviation
\item FIELD NAME: b_sname
\item FIELD FORMAT: Varchar, 6 Characters
\item DEFINITION: A unique institutionally given building abbreviation for the unique building name on your campus
}
}
\examples{
b_07()

}
