% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_08}
\alias{b_08}
\title{Calculate USHE Element b_08 (Year Building Constructed)}
\usage{
b_08(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_construction_year).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_08 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_08 (Year Building Constructed)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Year Building Constructed
\item FIELD NAME: b_year_cons
\item FIELD FORMAT: Varchar, 4 Characters
\item DEFINITION: Year Building Constructed.
}
}
\examples{
b_08()

}
