% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_09}
\alias{b_09}
\title{Calculate USHE Element b_09 (Year of Latest Major Remodeling)}
\usage{
b_09(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_remodel_year).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_09 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_09 (Year of Latest Major Remodeling)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Year of Latest Major Remodeling
\item FIELD NAME: b_year_rem
\item FIELD FORMAT: Varchar, 4 Characters
\item DEFINITION: The year of the most recent major renovation that brought the entire building to current codes and resulted in building systems (i.e. HVAC, electrical, plumbing, structural, envelope) being in good condition.
}
}
\examples{
b_09()

}
