% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_10}
\alias{b_10}
\title{Calculate USHE Element b_10 (Replacement Cost)}
\usage{
b_10(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_cost_replacement).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_10 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_10 (Replacement Cost)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Replacement Cost
\item FIELD NAME: b_replace_cost
\item FIELD FORMAT: Numeric (12,2) Characters. Two trailing 0’s added to reflect implied decimal places.
\item DEFINITION: The cost to replace the building as reported and supported by Risk Management. We need to have trailing "00"s.
}
}
\examples{
b_10()

}
