% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_11}
\alias{b_11}
\title{Calculate USHE Element b_11 (Building Condition)}
\usage{
b_11(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_condition_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_11 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_11 (Building Condition)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Condition
\item FIELD NAME: b_condition
\item FIELD FORMAT: Numeric 1 Character.
\item DEFINITION: Indicates the physical status of the building included in the inventory, 5 - 1 (good - unsatisfactory)
}
}
\examples{
b_11()

}
