% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_13}
\alias{b_13}
\title{Calculate USHE Element b_13 (Cost of Operating the Building)}
\usage{
b_13(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_cost_myr).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_13 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_13 (Cost of Operating the Building)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Cost of Operating the Building
\item FIELD NAME: b_cost_myr
\item FIELD FORMAT:  Numeric (8, 2) Characters.
\item DEFINITION:
}
}
\examples{
b_13()

}
