% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_14}
\alias{b_14}
\title{Calculate USHE Element b_14 (Building Risk Management Number)}
\usage{
b_14(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_number).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_14 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_14 (Building Risk Management Number)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Risk Management Number
\item FIELD NAME: b_rsknbr
\item FIELD FORMAT:  Numeric 5 Characters.
\item DEFINITION: A unique identifier assigned by State Risk Management for the building where the room is located. At some institutions, this number and B_NUMBER are the same value.
}
}
\examples{
b_14()

}
