% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{b_15}
\alias{b_15}
\title{Calculate USHE Element b_15 (Auxiliary Building)}
\usage{
b_15(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_auxiliary).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element b_15 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element b_15 (Auxiliary Building)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Auxiliary Building
\item FIELD NAME: b_aux
\item FIELD FORMAT:  Numeric 1 Character.
\item DEFINITION: Business space or other support facilities (as distinguished from primary programs of  instruction, research, and public service, and from organized activities and intercollegiate athletics)
}
}
\examples{
b_15()

}
