% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_03}
\alias{c_03}
\title{Calculate USHE Element c_03 (Course Subject)}
\usage{
c_03(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (subject_code).}
}
\value{
Original data frame, with USHE data element c_03 appended.
}
\description{
Calculate USHE Element c_03 (Course Subject)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_03()

}
