% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_04}
\alias{c_04}
\title{Calculate USHE Element c_04 (Course Number)}
\usage{
c_04(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_number).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_04 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_04 (Course Number)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_04()

}
