% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_05}
\alias{c_05}
\title{Calculate USHE Element c_05 (Course Section)}
\usage{
c_05(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (section_number).}
}
\value{
Original data frame, with USHE data element c_05 appended.
}
\description{
Calculate USHE Element c_05 (Course Section)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_05()

}
