% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_06}
\alias{c_06}
\title{Calculate USHE Element c_06 (Min Credit Hours)}
\usage{
c_06(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_min_credits).}
}
\value{
Original data frame, with USHE data element c_06 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_06 (Min Credit Hours)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_06()

}
