% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_07}
\alias{c_07}
\title{Calculate USHE Element c_07 (Max Credit Hours)}
\usage{
c_07(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_max_credits).}
}
\value{
Original data frame, with USHE data element c_07 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_07 (Max Credit Hours)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_07()

}
