% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_09}
\alias{c_09}
\title{Calculate USHE Element c_09 (Line Item)}
\usage{
c_09(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: ().}
}
\value{
Original data frame, with USHE data element c_09 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_09 (Line Item)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_09()

}
