% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_10}
\alias{c_10}
\title{Calculate USHE Element c_10 (Site Type)}
\usage{
c_10(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (campus_id).}
}
\value{
Original data frame, with USHE data element c_10 appended.
}
\description{
Calculate USHE Element c_10 (Site Type)
}
\details{
\strong{USHE Documentation}
ELEMENT NAME: Course Site Type
FIELD NAME: C_SITE_TYPE
FIELD FORMAT: Varchar, 3 Characters
DEFINITION: Code used to specify a site type where instruction originates.
}
\examples{
c_10()

}
