% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_12}
\alias{c_12}
\title{Calculate USHE Element c_12 (Delivery Method)}
\usage{
c_12(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (instruction_method_code).}
}
\value{
Original data frame, with USHE data element c_12 appended.
}
\description{
Calculate USHE Element c_12 (Delivery Method)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_12()

}
