% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_13}
\alias{c_13}
\title{Calculate USHE Element c_13 (Program Type)}
\usage{
c_13(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (program_type).}
}
\value{
Original data frame, with USHE data element c_13 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_13 (Program Type)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_13()

}
