% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_17}
\alias{c_17}
\title{Calculate USHE Element c_17 (Course Days 1)}
\usage{
c_17(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_days_1).}
}
\value{
Original data frame, with USHE data element c_17 appended.
}
\description{
Calculate USHE Element c_17 (Course Days 1)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_17()

}
