% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_19}
\alias{c_19}
\title{Calculate USHE Element c_19 (Building Number)}
\usage{
c_19(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_number_1).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_19 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_19 (Building Number)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_19()

}
