% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_20}
\alias{c_20}
\title{Calculate USHE Element c_20 (Room Number)}
\usage{
c_20(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_room_number_1).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_20 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_20 (Room Number)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_20()

}
