% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_21}
\alias{c_21}
\title{Calculate USHE Element c_21 (Maximum Room Occupancy)}
\usage{
c_21(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_max_occupancy_1).}
}
\value{
Original data frame, with USHE data element c_21 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_21 (Maximum Room Occupancy)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_21()

}
