% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_22}
\alias{c_22}
\title{Calculate USHE Element c_22 (Room Use Type)}
\usage{
c_22(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_use_code_1).}
}
\value{
Original data frame, with USHE data element c_22 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_22 (Room Use Type)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_22()

}
