% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_23}
\alias{c_23}
\title{Calculate USHE Element c_23 (Course Start Time 2)}
\usage{
c_23(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_start_time_2).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_23 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_23 (Course Start Time 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_23()

}
