% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_24}
\alias{c_24}
\title{Calculate USHE Element c_24 (Course Stop Time 2)}
\usage{
c_24(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_end_time_2).}
}
\value{
Original data frame, with USHE data element c_24 appended.
}
\description{
Calculate USHE Element c_24 (Course Stop Time 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_24()

}
