% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_25}
\alias{c_25}
\title{Calculate USHE Element c_25 (Course Days 2)}
\usage{
c_25(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_days_2).}
}
\value{
Original data frame, with USHE data element c_25 appended.
}
\description{
Calculate USHE Element c_25 (Course Days 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_25()

}
