% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_26}
\alias{c_26}
\title{Calculate USHE Element c_26 (Building Short Name 2)}
\usage{
c_26(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_building_id_2).}
}
\value{
Original data frame, with USHE data element c_26 appended.
}
\description{
Calculate USHE Element c_26 (Building Short Name 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_26()

}
