% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_28}
\alias{c_28}
\title{Calculate USHE Element c_28 (Room Number 2)}
\usage{
c_28(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_room_number_2).}
}
\value{
Original data frame, with USHE data element c_28 appended.
}
\description{
Calculate USHE Element c_28 (Room Number 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_28()

}
