% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_29}
\alias{c_29}
\title{Calculate USHE Element c_29 (Maximum Room Occupancy 2)}
\usage{
c_29(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_max_occupancy_2).}
}
\value{
Original data frame, with USHE data element c_29 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_29 (Maximum Room Occupancy 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_29()

}
