% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_31}
\alias{c_31}
\title{Calculate USHE Element c_31 (Course Start Time 3)}
\usage{
c_31(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_start_time_3).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_31 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_31 (Course Start Time 3)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_31()

}
