% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_34}
\alias{c_34}
\title{Calculate USHE Element c_34 (Building Short Name 3)}
\usage{
c_34(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_building_id_3).}
}
\value{
Original data frame, with USHE data element c_34 appended.
}
\description{
Calculate USHE Element c_34 (Building Short Name 3)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_34()

}
