% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_36}
\alias{c_36}
\title{Calculate USHE Element c_36 (Room Number 3)}
\usage{
c_36(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_room_number_3).}
}
\value{
Original data frame, with USHE data element c_36 appended.
}
\description{
Calculate USHE Element c_36 (Room Number 3)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_36()

}
