% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_38}
\alias{c_38}
\title{Calculate USHE Element c_38 (Room Use Type 3)}
\usage{
c_38(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_use_code_3).}
}
\value{
Original data frame, with USHE data element c_38 appended.
}
\description{
Calculate USHE Element c_38 (Room Use Type 3)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_38()

}
