% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_39}
\alias{c_39}
\title{Calculate USHE Element c_39 (Course Start Date)}
\usage{
c_39(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_start_date).}
}
\value{
Original data frame, with USHE data element c_39 appended.
}
\description{
Calculate USHE Element c_39 (Course Start Date)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_39()

}
