% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_41}
\alias{c_41}
\title{Calculate USHE Element c_41 (Course Title)}
\usage{
c_41(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_title).}
}
\value{
Original data frame, with USHE data element c_41 appended.
}
\description{
Calculate USHE Element c_41 (Course Title)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_41()

}
