% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_43}
\alias{c_43}
\title{Calculate USHE Element c_43 (Instructor Name)}
\usage{
c_43(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (instructor_first_name, instructor_last_name, instructor_middle_name).}
}
\value{
Original data frame, with USHE data element c_43 appended.
}
\description{
Calculate USHE Element c_43 (Instructor Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_43()

}
