% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_44}
\alias{c_44}
\title{Calculate USHE Element c_44 (Course Instruction Type)}
\usage{
c_44(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (section_format_type_code).}
}
\value{
Original data frame, with USHE data element c_44 appended.
}
\description{
Calculate USHE Element c_44 (Course Instruction Type)
}
\details{
\strong{USHE Documentation}
Element: C-44
ELEMENT NAME: Instructional Type
FIELD NAME: C_INSTRUCT_TYPE
FIELD FORMAT: Varchar, 3 Characters
}
\examples{
c_44()

}
