% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_46}
\alias{c_46}
\title{Calculate USHE Element c_46 (Course Department)}
\usage{
c_46(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (academic_department_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_46 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_46 (Course Department)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_46()

}
