% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_47}
\alias{c_47}
\title{Calculate USHE Element c_47 (Course General Education Code)}
\usage{
c_47(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (attribute_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_47 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_47 (Course General Education Code)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: General Education Course Codes
\item FIELD NAME: C_GEN_ED
\item FIELD FORMAT: Varchar, 4 Characters
\item DEFINITION: The codes which designate the required General Education requirements as outlined in Regent Policy R470.
Do NOT report courses beyond the Breadth General Education Requirements such as:  writing or quantitative intensive.
}
}
\examples{
c_47()

}
