% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_48}
\alias{c_48}
\title{Calculate USHE Element c_48 (Course Destination Site)}
\usage{
c_48(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (budget_code, campus_id, instruction_method_code, section_number).}
}
\value{
Original data frame, with USHE data element c_48 appended.
}
\description{
Calculate USHE Element c_48 (Course Destination Site)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Course Destination Site
\item FIELD NAME: C_DEST_SITE
\item FIELD FORMAT: Varchar, 6 Characters
\item DEFINITION: Designed for concurrent enrollment courses.
The High School or Special  Secondary School code which uniquely identifies the destination site of a concurrent enrollment course.
This code will only be used where the site may be clearly identified and left blank when not.
The codes for any secondary institution located within the United States can be found by accessing the URL (http://www.act.org/aap/regist/lookuphs.html).
Sample codes are listed below.
}
}
\examples{
c_48()

}
