% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_50}
\alias{c_50}
\title{Calculate USHE Element c_50 (Delivery Model)}
\usage{
c_50(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (instruction_method_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_50 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_50 (Delivery Model)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Delivery Model
\item FIELD NAME: C_ DELIVERY_MODEL
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: The model in which the course was taught.
}
}
\examples{
c_50()

}
