% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_51}
\alias{c_51}
\title{Calculate USHE Element c_51 (Class Level)}
\usage{
c_51(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_level_id, course_number, subject_code).}
}
\value{
Original data frame, with USHE data element c_51 appended.
}
\description{
Calculate USHE Element c_51 (Class Level)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Course Level
\item FIELD NAME: C_LEVEL
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: Indicates whether the course is offered as remedial or for undergraduate or  graduate level credit.
}
}
\examples{
c_51()

}
