% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_52}
\alias{c_52}
\title{Calculate USHE Element c_52 (Course Reference Number)}
\usage{
c_52(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_reference_number).}
}
\value{
Original data frame, with USHE data element c_52 appended.
}
\description{
Calculate USHE Element c_52 (Course Reference Number)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_52()

}
