% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_53}
\alias{c_53}
\title{Calculate USHE Element c_53 (Course Site Type 2)}
\usage{
c_53(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_building_id_2, campus_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_53 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_53 (Course Site Type 2)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_53()

}
