% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_54}
\alias{c_54}
\title{Calculate USHE Element c_54 (Course Site Type 3)}
\usage{
c_54(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_building_id_3, campus_id).}
}
\value{
Original data frame, with USHE data element c_54 appended.
}
\description{
Calculate USHE Element c_54 (Course Site Type 3)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_54()

}
