% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{clean}
\alias{clean}
\title{Clean Data Frame}
\usage{
clean(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame.}
}
\value{
Original data frame, with cleaning operations applied.
}
\description{
Performs various operations over each value in data frame.
}
\examples{
clean()

}
