% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{clean_days}
\alias{clean_days}
\title{Clean days of the week}
\usage{
clean_days(x)
}
\arguments{
\item{x}{a character element}
}
\value{
A clean name.
}
\description{
Clean days of the week
}
\details{
Removes extra spaces characters and truncation limit to 7 char
}
