% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{clean_name}
\alias{clean_name}
\title{Clean names}
\usage{
clean_name(x, .width = 15)
}
\arguments{
\item{x}{a character element}

\item{.width}{the truncation limit set to 15}
}
\value{
A clean name.
}
\description{
Clean names
}
\details{
Removes non-alpha characters and truncation limits in names
}
