% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws.R
\name{d_05}
\alias{d_05}
\title{Calculate USHE Element d_05 (End Search Date)}
\usage{
d_05(input_df = usheUtils::fake_dws_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (end_search_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element d_05 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element d_05 (End Search Date)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: End Search Date
\item FIELD NAME: d_end_date
\item FIELD FORMAT: Varchar, 8 Characters (YYYYMMDD)
\item DEFINITION: The calendar date of the End Search Date.
}
}
\examples{
d_05()

}
