% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws.R
\name{d_06}
\alias{d_06}
\title{Calculate USHE Element d_06 ( Description)}
\usage{
d_06(input_df = usheUtils::fake_dws_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (program_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element d_06 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element d_06 ( Description)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: E Description
\item FIELD NAME:  Description
\item FIELD FORMAT: Varchar, 100 Characters
\item DEFINITION: Institutions may use this field according to their preference or to differentiate students if submitting multiple files.
}
}
\examples{
d_06()

}
