% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_aid.R
\name{f_06}
\alias{f_06}
\title{Calculate USHE Element f_06 (Financial Aid Type)}
\usage{
f_06(input_df = usheUtils::fake_fin_aid_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (financial_aid_fund_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element f_06 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element f_06 (Financial Aid Type)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Financial Aid Type
\item FIELD NAME: f_type_code.
\item FIELD FORMAT: Varchar, 6 Characters
\item DEFINITION: The code that defines what type of financial aid the student was awarded.
Aid grouping is how aid is classified for both federal and state reporting.
}
}
\examples{
f_06()

}
