% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_aid.R
\name{f_07}
\alias{f_07}
\title{Calculate USHE Element f_07 (Financial Aid Amount)}
\usage{
f_07(input_df = usheUtils::fake_fin_aid_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (amount_offered).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element f_07 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element f_07 (Financial Aid Amount)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Financial Aid Amount
\item FIELD NAME: f_amount.
\item FIELD FORMAT: Numeric (8,2)
\item DEFINITION: Amount of Aid awarded to or Loans accepted by a student.
}
}
\examples{
f_07()

}
