% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_building_df.R
\docType{data}
\name{fake_building_df}
\alias{fake_building_df}
\title{Data of fake_building_df.R}
\format{
A data frame:
\describe{
\item{building_location_code}{character}
\item{building_ownership_code}{character}
\item{submission_year}{character}
\item{building_name}{character}
\item{building_number}{character}
\item{building_abbrv}{character}
\item{building_construction_year}{character}
\item{building_remodel_year}{character}
\item{building_cost_replacement}{character}
\item{building_condition_code}{character}
\item{building_area_gross}{character}
\item{building_cost_myr}{character}
\item{building_auxiliary}{character}
}
}
\usage{
fake_building_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
