% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_dws_df.R
\docType{data}
\name{fake_dws_df}
\alias{fake_dws_df}
\title{Data of fake_dws_df.R}
\format{
A data frame:
\describe{
\item{ssn}{character}
\item{student_id}{character}
\item{graduation_date}{character}
\item{end_search_date}{character}
\item{program_id}{character}
}
}
\usage{
fake_dws_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
