% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_din_aid_df.R
\docType{data}
\name{fake_fin_aid_df}
\alias{fake_fin_aid_df}
\title{Data of fake_fin_aid_df.R}
\format{
A data frame:
\describe{
\item{financial_aid_year_id}{character}
\item{term_id}{character}
\item{ssn}{character}
\item{season}{character}
\item{student_id}{character}
\item{financial_aid_fund_id}{character}
\item{amount_offered}{character}
}
}
\usage{
fake_fin_aid_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
