% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_mission_df.R
\docType{data}
\name{fake_mission_df}
\alias{fake_mission_df}
\title{Data of fake_mission_df.R}
\format{
A data frame:
\describe{
\item{student_id}{character}
\item{last_name}{character}
\item{first_name}{character}
\item{middle_name}{character}
\item{birth_date}{character}
\item{term_end_date}{character}
}
}
\usage{
fake_mission_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
