% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_program_df.R
\docType{data}
\name{fake_program_df}
\alias{fake_program_df}
\title{Data of fake_program_df.R}
\format{
A data frame:
\describe{
\item{major_desc}{character}
\item{required_credits}{character}
\item{is_perkins}{character}
\item{degree_id}{character}
\item{ipeds_award_level_code}{character}
\item{cip_code}{character}
\item{academic_year}{character}
}
}
\usage{
fake_program_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
