% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_rooms_df.R
\docType{data}
\name{fake_rooms_df}
\alias{fake_rooms_df}
\title{Data of fake_rooms_df.R}
\format{
A data frame:
\describe{
\item{submission_year}{character}
\item{building_number}{character}
\item{room_number}{character}
\item{room_group1_code}{character}
\item{room_use_code_group}{character}
\item{room_use_code}{character}
\item{room_name}{character}
\item{room_stations}{character}
\item{room_area}{character}
\item{room_disabled_access}{character}
\item{room_prorated}{character}
\item{room_prorated_area}{character}
\item{room_activity_date}{character}
}
}
\usage{
fake_rooms_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
