% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_student_course_df.R
\docType{data}
\name{fake_student_course_df}
\alias{fake_student_course_df}
\title{Data of fake_student_course_df.R}
\format{
A data frame:
\describe{
\item{subject_code}{character}
\item{course_number}{character}
\item{section_number}{character}
\item{attempted_credits}{character}
\item{earned_credits}{character}
\item{course_level_id}{character}
\item{final_grade}{character}
\item{student_type_code}{character}
\item{budget_code}{character}
\item{is_concurrent_course}{character}
\item{student_id}{character}
\item{ssn}{character}
\item{course_reference_number}{character}
\item{season}{character}
\item{academic_year_code}{character}
\item{version_id}{character}
}
}
\usage{
fake_student_course_df
}
\description{
A data frame containing ...
}
\keyword{datasets}
