% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_08}
\alias{g_08}
\title{Calculate USHE Element g_08 (Date of Graduation)}
\usage{
g_08(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (graduation_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_08 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_08 (Date of Graduation)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Date of Graduation
\item FIELD NAME: g_date
\item FIELD FORMAT: Varchar, 8 Characters (YYYYMMDD Format)
\item DEFINITION: The calendar date the formal award was conferred by the institution.
}
}
\examples{
g_08()

}
