% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_10}
\alias{g_10}
\title{Calculate USHE Element g_10 (Degree Type)}
\usage{
g_10(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (degree_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_10 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_10 (Degree Type)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Degree Type
\item FIELD NAME: g_deg_type
\item FIELD FORMAT: Varchar, 6 Characters
\item DEFINITION: The Level of Degree or Certificate Completed for the award conferred.
}
}
\examples{
g_10()

}
