% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_11}
\alias{g_11}
\title{Calculate USHE Element g_11 (Graduation GPA )}
\usage{
g_11(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (cumulative_graduation_gpa).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_11 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_11 (Graduation GPA )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Graduation GPA
\item FIELD NAME: g_gpa
\item FIELD FORMAT: Numeric (4,3)
\item DEFINITION: Student's overall cumulative GPA tied to their graduation award. All credit hours should represent average course grade on a 4.000 scale.
}
}
\examples{
g_11()

}
