% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_12}
\alias{g_12}
\title{Calculate USHE Element (Total Cum Transfer Hrs Accepted)}
\usage{
g_12(input_df = usheUtils::fake_graduation_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (transfer_cumulative_credits_earned, total_cumulative_ap_credits_earned, total_cumulative_clep_credits_earned ).}
}
\value{
Original data frame, with USHE data element g_12 appended.
}
\description{
Calculate USHE Element (Total Cum Transfer Hrs Accepted)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Cumulative Transfer Semester Credit Hours Accepted
\item FIELD NAME: G_TRANS_TOTAL
\item FIELD FORMAT: Numeric (5,1)
\item DEFINITION: Total number of credit hours accepted by graduation date at your institution relevant to the degree being awarded.
(e.g. Transfer credit from another institution). This does not include credits earned at your institution (see S-20).
This does not include AP, CLEP, Challenge, or Military  Credit. These hours are included in the G_HRS_OTHER field.
Hours should all be converted to  semester hours.
}
}
\examples{
g_12()

}
