% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_13}
\alias{g_13}
\title{Calculate USHE Element g_13 (Total Hours at Graduation )}
\usage{
g_13(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (overall_cumulative_credits_earned, total_remedial_hours).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_13 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_13 (Total Hours at Graduation )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Total Hours at Graduation
\item FIELD NAME: g_grad_hrs
\item FIELD FORMAT: Numeric (5,1)
\item DEFINITION: Total number of overall hours relevant to the degree awarded when the student graduated.
This field should include all college-level hours earned for undergraduate or graduate  coursework.
Remedial hours should be excluded. Hours should all be converted to semester
}
}
\examples{
g_13()

}
