% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_14}
\alias{g_14}
\title{Calculate USHE Element g_14 (Accepted Credit from Other Sources )}
\usage{
g_14(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (total_cumulative_credits_attempted_other_sources).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_14 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_14 (Accepted Credit from Other Sources )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Accepted Credit from Other Sources
\item FIELD NAME: g_hrs_other
\item FIELD FORMAT: Numeric (4,1)
\item DEFINITION: Hours from AP credit, CLEP test, Language test, Challenge, Military, etc. Hours  should all be converted to semester hours.
}
}
\examples{
g_14()

}
