% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_15}
\alias{g_15}
\title{Calculate USHE Element g_15 (Remedial Hours )}
\usage{
g_15(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (total_remedial_hours, level_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_15 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_15 (Remedial Hours )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Remedial Hours
\item FIELD NAME: g_remedial_hrs
\item FIELD FORMAT: Numeric (4,1)
\item DEFINITION: Remedial Hours which are included in S-20. Hours should all be converted to  semester hours. Applies only to undergrad degrees; grad degrees should be 0
}
}
\examples{
g_15()

}
