% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_16}
\alias{g_16}
\title{Calculate USHE Element g_16 (Previous Degree Type  )}
\usage{
g_16(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (previous_degree_type).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_16 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_16 (Previous Degree Type  )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Previous Degree Type
\item FIELD NAME: g_prev_deg_type
\item FIELD FORMAT: Varchar, 3 Characters
\item DEFINITION: Type of highest degree awarded earned prior to this reporting period at any  institution, including your own institution
}
}
\examples{
g_16()

}
