% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_17}
\alias{g_17}
\title{Calculate USHE Element g_17 (IPEDS Award Level Code  )}
\usage{
g_17(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ipeds_award_level_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_17 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_17 (IPEDS Award Level Code  )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: IPEDS Award Level Code
\item FIELD NAME: g_ipeds
\item FIELD FORMAT: Number, 2 Characters
\item DEFINITION: TThe Award Level codes that correspond with the IPEDS Completion Survey.
}
}
\examples{
g_17()

}
