% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_18}
\alias{g_18}
\title{Calculate USHE Element g_18 (Required Hours for Degree)}
\usage{
g_18(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (required_credits).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_18 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_18 (Required Hours for Degree)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Required Hours for Degree
\item FIELD NAME: g_req_hrs_deg
\item FIELD FORMAT: Numeric(3,0)
\item DEFINITION: The minimum number of credit hours required for the degree including any course pre requisites.
Not to include any major remedial hours. Hours should all be converted to semester hours.
}
}
\examples{
g_18()

}
