% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_19}
\alias{g_19}
\title{Calculate USHE Element g_19 (High School Codes)}
\usage{
g_19(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (high_school_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_19 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_19 (High School Codes)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: High School Codes
\item FIELD NAME: g_high_school
\item FIELD FORMAT: Varchar, 6 Characters
\item DEFINITION: The High School or Special Secondary School code which uniquely identifies each student’s secondary institution.
The codes for any secondary institution located within the United States can be found by accessing the URL
\link{act.org}(http://www.act.org/content/act/en/products-and services/the-act/registration/high-school-codes-lookup.html)
\strong{For undergraduate students only}
}
}
\examples{
g_19()

}
