% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_22}
\alias{g_22}
\title{Calculate USHE Element g_22 (Workforce Education Earned Contact Hours)}
\usage{
g_22(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (earned_contact_hrs).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_22 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_22 (Workforce Education Earned Contact Hours)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Workforce Education Earned Contact Hours
\item FIELD NAME: g_we_earned_contact_hrs
\item FIELD FORMAT: Numeric(5,1)
\item DEFINITION:The number of noncredit program hours the student earned in completion of the  Workforce Education Certificate program.
}
}
\examples{
g_22()

}
