% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_25}
\alias{g_25}
\title{Calculate USHE Element g_25 (Term of Graduation)}
\usage{
g_25(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (season).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_25 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_25 (Term of Graduation)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Term of Graduation
\item FIELD NAME: g_term
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: The term in which the student graduated.
}
}
\examples{
g_25()

}
