% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_26}
\alias{g_26}
\title{Calculate USHE Element g_26 (Major College)}
\usage{
g_26(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (college_desc).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_26 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_26 (Major College)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Major College
\item FIELD NAME: g_college
\item FIELD FORMAT: Varchar, 100 Characters
\item DEFINITION: The college or school that houses the student’s major. Should be tied to element G_09 CIP CODE.
}
}
\examples{
g_26()

}
