% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{g_27}
\alias{g_27}
\title{Calculate USHE Element g_27 (Major Name)}
\usage{
g_27(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (primary_major_desc).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_27 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_27 (Major Name)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Major Name
\item FIELD NAME: g_major
\item FIELD FORMAT: Varchar, 100 Characters
\item DEFINITION: Name of student’s major as listed on transcript. Should be tied to element G_09 CIP CODE.
}
}
\examples{
g_27()

}
