% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_birth_date}
\alias{gen_ushe_birth_date}
\alias{s_12}
\alias{m_03}
\alias{g_05}
\title{Calculate USHE Element (Birth Date)}
\usage{
gen_ushe_birth_date(input_df = usheUtils::fake_student_df)

s_12(input_df = usheUtils::fake_student_df)

m_03(input_df = usheUtils::fake_student_df)

g_05(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (birth_date).}
}
\value{
Original data frame, with USHE data elements gen_ushe_birth_date and m_03 appended.
}
\description{
Calculate USHE Element (Birth Date)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Date of Birth
\item FIELD NAME: S_BIRTH_DT, M_BIRTH_DT, & G_BIRTH_DT
\item FIELD FORMAT: Varchar, 8 Characters (YYYYMMDD),
\item DEFINITION: The calendar student date of birth, as designated by student.
}
}
\examples{
gen_ushe_birth_date()

s_12()
m_03()
g_05()
}
