% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_ethnicty}
\alias{gen_ushe_ethnicty}
\alias{s_14}
\alias{g_07}
\title{Calculate USHE Element Ethnicity (Ethnic Origin)}
\usage{
gen_ushe_ethnicty(input_df = usheUtils::fake_student_df)

s_14(input_df = usheUtils::fake_student_df)

g_07(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (is_hispanic_latino_ethnicity, is_asian, is_black, is_american_indian_alaskan, is_hawaiian_pacific_islander, is_white, is_international, is_other_race).}
}
\value{
Original data frame, with USHE data element gen_ushe_ethnicty appended.
}
\description{
Calculate USHE Element Ethnicity (Ethnic Origin)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Ethnicity
\item FIELD NAME: s_ethnic, g_ethnic
\item FIELD FORMAT: Varchar, 8 Characters,
\item DEFINITION: The racial and ethnic categories used to classify students.
}
}
\examples{
gen_ushe_ethnicty()

s_14()
g_07()
}
