% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_gender}
\alias{gen_ushe_gender}
\alias{s_13}
\alias{g_06}
\title{Calculate USHE Element(Gender)}
\usage{
gen_ushe_gender(input_df = usheUtils::fake_student_df)

s_13(input_df = usheUtils::fake_student_df)

g_06(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (gender_code).}
}
\value{
Original data frame, with USHE data element gen_ushe_gender appended.
}
\description{
Calculate USHE Element(Gender)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Gender
\item FIELD NAME: gen_ushe_gender, g_gender, s_gender
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: A code indicating the gender of the student
}
}
\examples{
gen_ushe_gender()

s_13()
g_06()
}
