% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_grad_date}
\alias{gen_ushe_grad_date}
\alias{g_08}
\alias{d_04}
\title{Calculate USHE Element g_08 (Date of Graduation)}
\usage{
gen_ushe_grad_date(
  input_df = usheUtils::fake_graduation_df,
  with_intermediates = FALSE
)

g_08(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)

d_04(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (graduation_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element g_08 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element g_08 (Date of Graduation)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Date of Graduation
\item FIELD NAME: gen_ushe_grad_date, g_date & d_start_dt
\item FIELD FORMAT: Varchar, 8 Characters (YYYYMMDD Format)
\item DEFINITION: The calendar date the formal award was conferred by the institution.
}
}
\examples{
gen_ushe_grad_date()

g_08()
d_04()
}
