% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_high_school}
\alias{gen_ushe_high_school}
\alias{g_19}
\alias{s_28}
\title{Calculate USHE Element (High School Codes)}
\usage{
gen_ushe_high_school(input_df = usheUtils::fake_student_df)

g_19(input_df = usheUtils::fake_student_df)

s_28(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (high_school_code).}
}
\value{
Original data frame, with USHE data element gen_ushe_high_school appended.
}
\description{
Calculate USHE Element (High School Codes)
}
\details{
\strong{USHE Documentation}
ELEMENT NAME: High School Codes
FIELD NAME: gen_ushe_high_school, g_high_school and s_high_school.
FIELD FORMAT: Varchar, 6 Characters
DEFINITION: The High School or Special Secondary School code which uniquely identifies each  student’s secondary institution.
}
\examples{
gen_ushe_high_school()

g_19()
s_28()
}
