% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_id}
\alias{gen_ushe_id}
\alias{s_03}
\alias{sc_03}
\alias{g_02}
\title{Calculate USHE Element Social Security Number or Institutionally Assigned ID (Student ID)}
\usage{
gen_ushe_id(input_df = usheUtils::fake_student_df)

s_03(input_df = usheUtils::fake_student_df)

sc_03(input_df = usheUtils::fake_student_df)

g_02(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (student_id, ssn).}
}
\value{
Original data frame, with USHE data element s_03 appended.
}
\description{
Calculate USHE Element Social Security Number or Institutionally Assigned ID (Student ID)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Institutionally Assigned ID
\item FIELD NAME: s_id, sc_id, and g_id
\item FIELD FORMAT: Varchar, 9 Characters
\item DEFINITION: The unique institutionally assigned identification number for each student intended to be used in lieu of using a student’s social security number.
G_Banner_ID must begin with an Alpha character representing the institution.
}
}
\examples{
gen_ushe_id()

s_03()
sc_03()
g_02()
}
