% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_inst}
\alias{gen_ushe_inst}
\alias{s_01}
\alias{c_01}
\alias{sc_01}
\alias{pf_01}
\alias{m_01}
\alias{b_01}
\alias{r_01}
\alias{g_01}
\title{Calculate USHE Element Institution (Institution)}
\usage{
gen_ushe_inst(input_df = usheUtils::fake_student_df)

s_01(input_df = usheUtils::fake_student_df)

c_01(input_df = usheUtils::fake_student_df)

sc_01(input_df = usheUtils::fake_student_df)

pf_01(input_df = usheUtils::fake_student_df)

m_01(input_df = usheUtils::fake_student_df)

b_01(input_df = usheUtils::fake_student_df)

r_01(input_df = usheUtils::fake_student_df)

g_01(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (no data fields required).}
}
\value{
Original data frame, with USHE data element s_01 appended.
}
\description{
Calculate USHE Element Institution (Institution)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Institution
\item FIELD NAME: C_INST, S_INST, G_INST, SC_INST, M_INST, F_INST, R_INST, B_INST, & PF_INST
\item FIELD FORMAT: Varchar, 4 Characters
\item DEFINITION: The identification number used by the U.S. Department of Education’s Office of Postsecondary Education (OPE) to identify schools that have Program Participation Agreements (PPA) so that its students are eligible to participate in Federal Student Financial Assistance programs under Title IV regulations.
This is a 6-digit number followed by a 2-digit suffix used to identify branches,  additional locations, and other entities that are part of the eligible institution.
USHE will use the last 4 digits of the 6-digit number for the field value as additional locations will be specified by C-09 Line Item.
}
}
\examples{
gen_ushe_inst()

s_01()
c_01()
sc_01()
pf_01()
m_01()
b_01()
r_01()
g_01()
}
