% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_ipeds}
\alias{gen_ushe_ipeds}
\alias{s_19}
\alias{pf_04}
\alias{g_17}
\title{Calculate USHE Element (IPEDS Award Level Code  )}
\usage{
gen_ushe_ipeds(
  input_df = usheUtils::fake_graduation_df,
  with_intermediates = FALSE
)

s_19(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)

pf_04(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)

g_17(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ipeds_award_level_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element gen_ushe_ipeds appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element (IPEDS Award Level Code  )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: IPEDS Award Level Code / Degree Level (for the programs file).
\item FIELD NAME: gen_ushe_ipeds, s_ipeds, pf_deg_level & g_ipeds
\item FIELD FORMAT: Number, 2 Characters
\item DEFINITION: The Award Level codes that correspond with the IPEDS Completion Survey.
}
}
\examples{
gen_ushe_ipeds()

s_19()
pf_04()
g_17()
}
