% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_name}
\alias{gen_ushe_name}
\alias{s_06}
\alias{g_03}
\title{Calculate USHE Elements for Names (Student Name)}
\usage{
gen_ushe_name(input_df = usheUtils::fake_student_df)

s_06(input_df = usheUtils::fake_student_df)

g_03(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (last_name, first_name, middle_name, name_suffix).}
}
\value{
Original data frame, with USHE data element s_06 appended.
}
\description{
Calculate USHE Elements for Names (Student Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
gen_ushe_name()

s_06()
g_03()
}
