% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_ssid}
\alias{gen_ushe_ssid}
\alias{s_34}
\alias{g_20}
\title{Calculate USHE Element s_34, g_20 (SSID aka USOE Unique ID)}
\usage{
gen_ushe_ssid(input_df = usheUtils::fake_student_df)

s_34(input_df = usheUtils::fake_student_df)

g_20(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ssid).}
}
\value{
Original data frame, with USHE data element gen_ushe_ssid appended.
}
\description{
Calculate USHE Element s_34, g_20 (SSID aka USOE Unique ID)
}
\details{
\strong{USHE Documentation}
ELEMENT NAME: Unique USOE State Student Identifier
FIELD NAME: gen_ushe_ssid, s_ssid & s_G_SSID
FIELD FORMAT: Varchar, 9 Characters
DEFINITION: The unique Utah State Student Identifier as assigned to each Utah public education student by the Utah State Board of Education.
This 9-digit number will appear on all high school transcripts beginning the 2006-2007 academic year.
}
\examples{
gen_ushe_ssid()

s_34()
g_20()
}
