% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_term}
\alias{gen_ushe_term}
\alias{g_25}
\alias{f_03}
\title{Calculate USHE Element (Term)}
\usage{
gen_ushe_term(
  input_df = usheUtils::fake_graduation_df,
  with_intermediates = FALSE
)

g_25(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)

f_03(input_df = usheUtils::fake_graduation_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (season).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element gen_ushe_ipeds appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element (Term)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Term
\item FIELD NAME: gen_ushe_term, g_term, f_term.
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: The term in which the student is evolved in either gradution or financial aid.
}
}
\examples{
gen_ushe_term()

g_25()
f_03()
}
