% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{gen_ushe_year_term_extract}
\alias{gen_ushe_year_term_extract}
\alias{s_02}
\alias{c_02}
\alias{sc_02}
\title{Calculate USHE Element (Year, Term, & Extract)}
\usage{
gen_ushe_year_term_extract(input_df = usheUtils::fake_student_df)

s_02(input_df = usheUtils::fake_student_df)

c_02(input_df = usheUtils::fake_student_df)

sc_02(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (season, academic_year_code, version_id).}
}
\value{
Original data frame, with USHE data element gen_ushe_year_term_extract s_02, sc_02, c_02 appended.
}
\description{
Calculate USHE Element (Year, Term, & Extract)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Year, Term & Extract Code
\item FIELD NAME: gen_ushe_, s_, sc_term_id & c_term_id
\item FIELD FORMAT: Number, 4 Characters (YYYY format)
\item Number, 1 Character (T format)
\item Varchar, 1 character (E format)
\item DEFINITION: The current academic year, term, and extract in which course is offered.
Format YYYYTE is made up of academic year YYYY, academic term T, and extract E (3rd week or End of Term).
}
}
\examples{
gen_ushe_year_term_extract()

s_02()
c_02()
sc_02()
}
