% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{generate_course_submission_file}
\alias{generate_course_submission_file}
\title{Generate Course Submission File}
\usage{
generate_course_submission_file(
  input_df = usheUtils::fake_course_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_reference_number,
course_number,
subject_code,
class_size,
budget_code,
campus_id,
section_number,
attribute_code,
academic_department_id,
college_id,
section_format_type_code,
instructor_first_name,
instructor_last_name,
instructor_employee_id,
course_title,
meet_end_date,
meet_start_date,
room_use_code_3,
room_max_occupancy_3,
meet_room_number_3,
building_number_3,
meet_building_id_3,
meet_days_3,
meet_end_time_3,
meet_start_time_3,
room_use_code_2,
room_max_occupancy_2,
meet_room_number_2,
building_number_2,
meet_building_id_2,
meet_days_2,
meet_end_time_2,
meet_start_time_2,
room_use_code_1,
room_max_occupancy_1,
meet_room_number_1,
building_number_1,
meet_building_id_1,
meet_days_1,
meet_end_time_1,
meet_start_time_1,
course_level_id,
program_type,
instruction_method_code,
contact_hours,
course_max_credits,
course_min_credits,
season,
academic_year_code,
version_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Course Submission File
}
\examples{
generate_course_submission_file()

}
