% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws.R
\name{generate_dws_submission_file}
\alias{generate_dws_submission_file}
\title{Generate DWS Submission File}
\usage{
generate_dws_submission_file(
  input_df = usheUtils::fake_dws_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ssn,
student_id,
graduation_date,
end_search_date,
program_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate DWS Submission File
}
\examples{
generate_dws_submission_file()

}
