% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_aid.R
\name{generate_financial_aid_submission_file}
\alias{generate_financial_aid_submission_file}
\title{Generate Financial Aid Submission File}
\usage{
generate_financial_aid_submission_file(
  input_df = usheUtils::fake_fin_aid_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (financial_aid_year_id,
term_id,
season,
ssn,
student_id,
financial_aid_fund_id,
amount_offered).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Financial Aid Submission File
}
\examples{
generate_financial_aid_submission_file()

}
