% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{generate_graduation_submission_file}
\alias{generate_graduation_submission_file}
\title{Generate Graduation Submission File}
\usage{
generate_graduation_submission_file(
  input_df = usheUtils::fake_graduation_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (student_id,
ssn,
last_name,
first_name,
middle_name,
name_suffix,
first_admit_state_code,
first_admit_county_code,
first_admit_country_iso_code,
birth_date,
gender_code,
is_hispanic_latino_ethnicity,
is_asian,
is_black,
is_american_indian_alaskan,
is_hawaiian_pacific_islander,
is_white,
is_international,
is_other_race,
graduation_date,
primary_major_cip_code,
degree_id,
degree_status_code,
cumulative_graduation_gpa,
transfer_cumulative_credits_earned,
total_cumulative_ap_credits_earned,
total_cumulative_clep_credits_earned,
cumulative_credits_earned,
total_remedial_hours,
total_cumulative_credits_attempted_other_sources,
level_id,
previous_degree_type,
ipeds_award_level_code,
required_credits,
high_school_code,
ssid,
earned_contact_hrs ,
program_hrs,
graduated_term_id,
financial_aid_year_id,
season,
primary_major_college_desc,
primary_major_desc,
degree_desc)}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Graduation Submission File
}
\examples{
generate_graduation_submission_file()

}
