% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missionary.R
\name{generate_missionary_submission_file}
\alias{generate_missionary_submission_file}
\title{Generate Missionary Submission File}
\usage{
generate_missionary_submission_file(
  input_df = usheUtils::fake_mission_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (last_name,
first_name,
middle_name,
birth_date,
term_end_date,
student_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Missionary Submission File
}
\examples{
generate_missionary_submission_file()

}
