% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{generate_program_validation_file}
\alias{generate_program_validation_file}
\title{Generate Programs Validation File}
\usage{
generate_program_validation_file(input_df = usheUtils::fake_program_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (major_desc,
required_credits,
is_perkins,
degree_id,
ipeds_award_level_code,
cip_code,
academic_year).}
}
\value{
A Data Frame, with all of the intermediate values used to create the USHE elements required for upload submission.
}
\description{
Generate Programs Validation File
}
\examples{
generate_program_validation_file()

}
