% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\name{generate_space_inventory_building_submission_file}
\alias{generate_space_inventory_building_submission_file}
\title{Generate Space Inventory Building Submission File}
\usage{
generate_space_inventory_building_submission_file(
  input_df = usheUtils::fake_building_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_location_code,
ownership_code,
submission_year,
building_name,
building_number,
building_abbrv,
building_construction_year,
building_remodel_year,
building_cost_replacement,
building_condition_code,
building_area_gross,
building_cost_myr,
building_number,
building_auxiliary).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Space Inventory Building Submission File
}
\examples{
generate_space_inventory_building_submission_file()

}
