% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{generate_space_inventory_rooms_submission_file}
\alias{generate_space_inventory_rooms_submission_file}
\title{Generate Space Inventory Rooms Submission File}
\usage{
generate_space_inventory_rooms_submission_file(
  input_df = usheUtils::fake_rooms_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (season,
academic_year_code,
version_id,
building_id,
room_number,
room_group1_code,
room_use_code_group,
room_use_code,
room_name,
room_stations,
room_area,
room_disabled_access,
room_prorated,
room_prorated_area,
room_activity_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Space Inventory Rooms Submission File
}
\examples{
generate_space_inventory_rooms_submission_file()

}
