% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{generate_student_validation_file}
\alias{generate_student_validation_file}
\title{Generate Student Validation File}
\usage{
generate_student_validation_file(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (list, of, required, data, fields).}
}
\value{
A Data Frame, with all of the intermediate values used to create the USHE elements required for upload submission.
}
\description{
Generate Student Validation File
}
\examples{
generate_student_validation_file()

}
