% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{is_valid_ssn}
\alias{is_valid_ssn}
\title{Check for valid SSN characters}
\usage{
is_valid_ssn(ssn)
}
\arguments{
\item{ssn}{a vector of ssn to check for validity.}
}
\value{
A Boolean TRUE or FALSE if SSN is valid.
}
\description{
Check for valid SSN characters
}
\details{
Is valid ssn id
}
\examples{
is_valid_ssn(c("123-45-6789", "987-65-4321"))

}
