% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missionary.R
\name{m_04}
\alias{m_04}
\title{Calculate USHE Element m_04 (Start Search Date)}
\usage{
m_04(input_df = usheUtils::fake_mission_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (term_end_date).}
}
\value{
Original data frame, with USHE data element m_04 appended.
}
\description{
Calculate USHE Element m_04 (Start Search Date)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME:
\item FIELD NAME: m_end_date
\item FIELD FORMAT:
\item DEFINITION:
}
}
\examples{
m_04()

}
