% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missionary.R
\name{m_05}
\alias{m_05}
\title{Calculate USHE Element m_05 (End Search Date)}
\usage{
m_05(input_df = usheUtils::fake_mission_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: ().}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element m_05 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element m_05 (End Search Date)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME:
\item FIELD NAME: m_start_date
\item FIELD FORMAT:
\item DEFINITION:
}
}
\examples{
m_05()

}
