% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_02}
\alias{pf_02}
\title{Calculate USHE Element pf_02 (Academic Year)}
\usage{
pf_02(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (academic_year).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_02 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_02 (Academic Year)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_02()

}
