% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_03}
\alias{pf_03}
\title{Calculate USHE Element pf_03 (CIP Code)}
\usage{
pf_03(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (cip_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_03 (CIP Code)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_03()

}
