% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_06}
\alias{pf_06}
\title{Calculate USHE Element pf_06 (Perkins)}
\usage{
pf_06(input_df = usheUtils::fake_program_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (is_perkins).}
}
\value{
Original data frame, with USHE data element pf_06 appended.
}
\description{
Calculate USHE Element pf_06 (Perkins)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_06()

}
