% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_07}
\alias{pf_07}
\title{Calculate USHE Element pf_07 (Technical Education)}
\usage{
pf_07(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: ().}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_07 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_07 (Technical Education)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_07()

}
