% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_09}
\alias{pf_09}
\title{Calculate USHE Element pf_09 (Program Participation Agreement)}
\usage{
pf_09(input_df = usheUtils::fake_program_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: ().}
}
\value{
Original data frame, with USHE data element pf_09 appended.
}
\description{
Calculate USHE Element pf_09 (Program Participation Agreement)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_09()

}
