% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_02}
\alias{r_02}
\title{Calculate USHE Element r_02 (Year Submission)}
\usage{
r_02(input_df = usheUtils::fake_building_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (submission_year).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_02 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_02 (Year Submission)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Submission Year
\item FIELD NAME: r_year
\item FIELD FORMAT: Varchar, 4 Characters (YYYY format)
\item DEFINITION:The fiscal year in which the Space Inventory - Building inventory was conducted.
}
}
\examples{
r_02()

}
