% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_03}
\alias{r_03}
\title{Calculate USHE Element r_03 (Building Number )}
\usage{
r_03(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (building_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_03 (Building Number )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Building Number
\item FIELD NAME: r_build_number
\item FIELD FORMAT: Varchar, 7 Characters
\item DEFINITION: A unique identifier assigned by the institution for the building where the room  is located. For rooms assigned as classroom space, this should match C_BLDG_NUM, C_BLDG_NUM2, or C_BLDG_NUM3 in the Courses file for student enrollments.
}
}
\examples{
r_03()

}
