% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_04}
\alias{r_04}
\title{Calculate USHE Element r_04 (Room Number )}
\usage{
r_04(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_number).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_04 (Room Number )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Room Number
\item FIELD NAME: r_number
\item FIELD FORMAT: Varchar, 30 Characters
\item DEFINITION: A unique institutionally given room number which will assist the institution in determining the location and possible use of a room.
}
}
\examples{
r_04()

}
