% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_05}
\alias{r_05}
\title{Calculate USHE Element r_05 (Room Suffix )}
\usage{
r_05(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: ().}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_05 (Room Suffix )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Room Suffix
\item FIELD NAME: r_suffix
\item FIELD FORMAT: Varchar, 10 Characters
\item DEFINITION: A unique institutionally given suffix to room’s coding structure to define subset rooms in a suite or defined area of use.
}
}
\examples{
r_05()

}
