% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_06}
\alias{r_06}
\title{Calculate USHE Element r_06 (Room Grouping Codes)}
\usage{
r_06(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_group1_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_06 (Room Grouping Codes)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Room Grouping Codes
\item FIELD NAME: r_group1
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: These space categories have been established to assist directly with the CDP Process of capital facilities needs in the Utah System of Higher Education.
}
}
\examples{
r_06()

}
