% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_07}
\alias{r_07}
\title{Calculate USHE Element r_07 (Category of Room Use )}
\usage{
r_07(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_use_code_group).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_07 (Category of Room Use )
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Category of room use
\item FIELD NAME: r_use_category
\item FIELD FORMAT: Varchar, 3 Characters
\item DEFINITION: This field identifies which category of use a room is assigned.
}
}
\examples{
r_07()

}
