% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_08}
\alias{r_08}
\title{Calculate USHE Element r_08 (Room Use Codes)}
\usage{
r_08(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_use_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_08 (Room Use Codes)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Room Use Codes
\item FIELD NAME: r_use_code
\item FIELD FORMAT: Varchar, 3 Characters
\item DEFINITION: Room use classification codes. Service codes are used for associated support rooms.
}
}
\examples{
r_08()

}
