% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_10}
\alias{r_10}
\title{Calculate USHE Element r_10 (Number of Seats or Stations in the Space)}
\usage{
r_10(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_stations).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_10 (Number of Seats or Stations in the Space)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Number of Seats or Stations in the Space
\item FIELD NAME: r_stations
\item FIELD FORMAT:  Numeric (5, 0) Characters,
\item DEFINITION:  Identified the capacity of the room for selected room use categories where  information about capacity (number of workstations, seats, or beds, for example)
}
}
\examples{
r_10()

}
