% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_11}
\alias{r_11}
\title{Calculate USHE Element r_11 (Assignable Area of the Room)}
\usage{
r_11(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_area).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_11 (Assignable Area of the Room)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Assignable Area of the Room
\item FIELD NAME: r_area
\item FIELD FORMAT:  Numeric (10, 0) Characters,
\item DEFINITION:The assignable floor area of the room, measured in assignable square feet. This is the total floor area of the room available
}
}
\examples{
r_11()

}
