% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_14}
\alias{r_14}
\title{Calculate USHE Element r_14 (Room’s Prorated Area)}
\usage{
r_14(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_prorated_area).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_14 (Room’s Prorated Area)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Room’s Prorated Area
\item FIELD NAME: r_prorated_area
\item FIELD FORMAT: Numeric, (6, 0) Characters
\item DEFINITION: Indicates the room’s prorated square footage.
}
}
\examples{
r_14()

}
