% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{r_15}
\alias{r_15}
\title{Calculate USHE Element r_15 (Data Updated Date)}
\usage{
r_15(input_df = usheUtils::fake_rooms_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (room_activity_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element r_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element r_15 (Data Updated Date)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Data Updated Date
\item FIELD NAME: r_update_date
\item FIELD FORMAT: Varchar, 8 Characters (YYYYMMDD)
\item DEFINITION: he date in which any of the data for the particular room was updated.
}
}
\examples{
r_15()

}
