% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{remove_special_characters}
\alias{remove_special_characters}
\title{Remove Special Characters}
\usage{
remove_special_characters(x = "spe;c.l-cha,ac?ers")
}
\arguments{
\item{x}{any value or vector of values}
}
\value{
x with special characters removed from x
}
\description{
Remove Special Characters
}
\examples{
remove_special_characters()
}
