% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_01}
\alias{s_01}
\title{Append Column s_01 (Institution)}
\usage{
s_01(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element s_01 appended.
}
\description{
Append Column s_01 (Institution)
}
\examples{
s_01()

}
