% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_02.R
\name{s_02}
\alias{s_02}
\title{Append Column s_02 (Year, Term, & Extract)}
\usage{
s_02(input_df = fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element s_02 appended.
}
\description{
Append Column s_02 (Year, Term, & Extract)
}
\examples{
s_02()

}
