% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_03}
\alias{s_03}
\title{Calculate USHE Element s_03 (Student ID)}
\usage{
s_03(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (student_id, student_ssn).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_03 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_03 (Student ID)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Institutionally Assigned ID
\item FIELD NAME: G_BANNER_ID
\item FIELD FORMAT: Varchar, 9 Characters
\item DEFINITION: The unique institutionally assigned identification number for each student intended to be used in lieu of using a student’s social security number.
G_Banner_ID must begin with an Alpha character representing the institution.
}
}
\examples{
s_03()

}
