% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_05}
\alias{s_05}
\title{Calculate USHE Element s_05 (Previous ID)}
\usage{
s_05(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (previous_student_id).}
}
\value{
Original data frame, with USHE data element s_05 appended.
}
\description{
Calculate USHE Element s_05 (Previous ID)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_05()

}
