% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_07}
\alias{s_07}
\title{Calculate USHE Element s_07 (Student Previous Name)}
\usage{
s_07(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (previous_last_name, previous_first_name, previous_middle_name, previous_name_suffix).}
}
\value{
Original data frame, with USHE data element s_07 appended.
}
\description{
Calculate USHE Element s_07 (Student Previous Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_07()

}
