% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_08}
\alias{s_08}
\title{Calculate USHE Element s_08 (Zip Code)}
\usage{
s_08(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (local_address_zip_code, mailing_address_zip_code).}
}
\value{
Original data frame, with USHE data element s_08 appended.
}
\description{
Calculate USHE Element s_08 (Zip Code)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Current Zip Code
\item FIELD NAME: S_CURR_ZIP
\item FIELD FORMAT: Varchar, 10 Characters
\item DEFINITION: The postal code of the student's current local address while attending classes.
}

Using Regex to check that zip code is in a valid US zip code format.
More information can be found: https://regexlib.com/Search.aspx?k=us+zip+code
}
\examples{
s_08()

}
