% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_11}
\alias{s_11}
\title{Calculate USHE Element s_11 (State Origin Code)}
\usage{
s_11(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (first_admit_state_code, first_admit_country_iso_code).}
}
\value{
Original data frame, with USHE data element s_11 appended.
}
\description{
Calculate USHE Element s_11 (State Origin Code)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_11()

}
