% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_12}
\alias{s_12}
\title{Append Column s_12 (Birth Date)}
\usage{
s_12(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element s_12 appended.
}
\description{
Append Column s_12 (Birth Date)
}
\examples{
s_12()

}
