% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_13}
\alias{s_13}
\title{Calculate USHE Element s_13 (Gender)}
\usage{
s_13(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (gender_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_13 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_13 (Gender)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_13()

}
