% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_14}
\alias{s_14}
\title{Append Column S14 (Ethnic Origin)}
\usage{
s_14(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{}
}
\value{
Original data frame, with USHE data Element s_14 appended.
}
\description{
Append Column S14 (Ethnic Origin)
}
\examples{
s_14()

}
