% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_15}
\alias{s_15}
\title{Calculate USHE Element s_15 (Residency)}
\usage{
s_15(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (residency_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_15 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_15 (Residency)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Regent Residency Status
\item FIELD NAME: S_REGENT_RES
\item FIELD FORMAT: Varchar, 1 Character,
}
}
\examples{
s_15()

}
