% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_16}
\alias{s_16}
\title{Calculate USHE Element s_16 (CIP Code)}
\usage{
s_16(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (primary_major_cip_code).}
}
\value{
Original data frame, with USHE data element s_16 appended.
}
\description{
Calculate USHE Element s_16 (CIP Code)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_16()

}
