% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_17}
\alias{s_17}
\title{Calculate USHE Element s_17 (Registration Status)}
\usage{
s_17(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (student_type_code, high_school_graduation_date, term_start_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_17 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_17 (Registration Status)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Registration Status
\item FIELD NAME: S_REG_STATUS
\item FIELD FORMAT: Varchar, 2 Characters,
\item DEFINITION: Student's entry status at the beginning of the term.
}
}
\examples{
s_17()

}
