% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_18}
\alias{s_18}
\title{Calculate USHE Element s_18 (Class Level)}
\usage{
s_18(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (primary_level_class_id).}
}
\value{
Original data frame, with USHE data element s_18 appended.
}
\description{
Calculate USHE Element s_18 (Class Level)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_18()

}
