% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_19}
\alias{s_19}
\title{Calculate USHE Element s_19 (Degree Intent)}
\usage{
s_19(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (primary_degree_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_19 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_19 (Degree Intent)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Degree Intent
\item FIELD NAME: S_DEG_INTENT
\item FIELD FORMAT: Varchar, 2 Character
\item DEFINITION: The length/nature of the degree that the student is working toward.
S_DEG_INTENT with S_CURR_CIP represents the student’s major and ties to S_MAJOR
}
}
\examples{
s_19()

}
