% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_20}
\alias{s_20}
\title{Calculate USHE Element s_20 (Total Cumulative Undergrad Hrs)}
\usage{
s_20(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (level_id, institutional_cumulative_credits_earned).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_20 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_20 (Total Cumulative Undergrad Hrs)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Cumulative Institutional Undergraduate Hours
\item FIELD NAME: S_CUM_HRS_UGRAD
\item FIELD FORMAT: Numeric(5,1)
\item DEFINITION: Total number of credit hours the student has earned as an undergraduate student at this institution.
Hours should be semester hours. Value should be zero for graduate students (see element S 18).
(Does not include transfer hours.)
}
}
\examples{
s_20()

}
