% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_21}
\alias{s_21}
\title{Calculate USHE Element s_21 (Cumulative Inst Undergrad GPA)}
\usage{
s_21(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (level_id, institutional_cumulative_gpa).}
}
\value{
Original data frame, with USHE data element s_21 appended.
}
\description{
Calculate USHE Element s_21 (Cumulative Inst Undergrad GPA)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Cumulative Institutional Undergraduate GPA
\item FIELD NAME: S_CUM_GPA_UGRAD
\item FIELD FORMAT: Numeric (4,3)
\item DEFINITION: Student's Cumulative GPA as of the present academic term as an undergraduate student.
This should tie to Cumulative Institutional Hours field.
No transfer hours are to be included.
All credit hours should represent average course grade on a 4.0 scale.
This GPA ties to earned credits in  Cumulative Institutional Undergraduate Hours.
Value should be zero for graduate students (see element  S-18).
}
}
\examples{
s_21()

}
