% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_24}
\alias{s_24}
\title{Calculate USHE Element s_24 (Total Cum U-grad Transfer Hrs Accepted)}
\usage{
s_24(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (transfer_cumulative_credits_earned, total_cumulative_clep_credits_earned, total_cumulative_ap_credits_earned).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_24 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_24 (Total Cum U-grad Transfer Hrs Accepted)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Cumulative Undergraduate Transfer Accepted
\item FIELD NAME: S_TRANS_TOTAL
\item FIELD FORMAT: Numeric (5,1)
\item DEFINITION: Total number of credit hours accepted to date at your institution (e.g. transfer credit from another institution).
Applies to undergraduate and graduate hours.
This includes credit from other institutions, Challenge, Military, and other test credit except CLEP and AP (see S-32 and S-33).
This does not include credits earned at your institution (see S-20).
Hours should all be converted to semester hours.
}
}
\examples{
s_24()

}
