% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_25}
\alias{s_25}
\title{Calculate USHE Element s_25 (Part-time/Full-time)}
\usage{
s_25(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (full_time_part_time_code).}
}
\value{
Original data frame, with USHE data element s_25 appended.
}
\description{
Calculate USHE Element s_25 (Part-time/Full-time)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_25()

}
