% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_26}
\alias{s_26}
\title{Calculate USHE Element s_26 (Age)}
\usage{
s_26(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (version_date, birth_date).}
}
\value{
Original data frame, with USHE data element s_26 appended.
}
\description{
Calculate USHE Element s_26 (Age)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Age at 3rd Week
\item FIELD NAME: S_AGE
\item FIELD FORMAT: Numeric (3,0)
\item DEFINITION: The age of the student at third week extract date.
}
}
\examples{
s_26()

}
