% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_27}
\alias{s_27}
\title{Calculate USHE Element s_27 (Country Origin Codes)}
\usage{
s_27(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (first_admit_country_iso_code).}
}
\value{
Original data frame, with USHE data element s_27 appended.
}
\description{
Calculate USHE Element s_27 (Country Origin Codes)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_27()

}
