% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_29}
\alias{s_29}
\title{Calculate USHE Element s_29 (Non Resident Tuition Waiver)}
\usage{
s_29(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (house_bill_75_waiver).}
}
\value{
Original data frame, with USHE data element s_29 appended.
}
\description{
Calculate USHE Element s_29 (Non Resident Tuition Waiver)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_29()

}
