% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_32}
\alias{s_32}
\title{Calculate USHE Element s_32 (Total Cum U-grad CLEP Cr Accepted)}
\usage{
s_32(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (total_cumulative_clep_credits_earned).}
}
\value{
Original data frame, with USHE data element s_32 appended.
}
\description{
Calculate USHE Element s_32 (Total Cum U-grad CLEP Cr Accepted)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_32()

}
