% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_33}
\alias{s_33}
\title{Calculate USHE Element s_33 (Total Cum U-grad AP Cr Accepted)}
\usage{
s_33(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (total_cumulative_ap_credits_earned).}
}
\value{
Original data frame, with USHE data element s_33 appended.
}
\description{
Calculate USHE Element s_33 (Total Cum U-grad AP Cr Accepted)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_33()

}
