% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_34}
\alias{s_34}
\title{Calculate USHE Element s_34 (SSID aka USOE Unique ID)}
\usage{
s_34(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ssid).}
}
\value{
Original data frame, with USHE data element s_34 appended.
}
\description{
Calculate USHE Element s_34 (SSID aka USOE Unique ID)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_34()

}
