% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_42}
\alias{s_42}
\title{Calculate USHE Element s_42 (High School Graduating Date)}
\usage{
s_42(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (high_school_graduation_date).}
}
\value{
Original data frame, with USHE data element s_42 appended.
}
\description{
Calculate USHE Element s_42 (High School Graduating Date)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_42()

}
