% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_43}
\alias{s_43}
\title{Calculate USHE Element s_43 (Term GPA)}
\usage{
s_43(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (institutional_term_gpa).}
}
\value{
Original data frame, with USHE data element s_43 appended.
}
\description{
Calculate USHE Element s_43 (Term GPA)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_43()

}
