% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_46}
\alias{s_46}
\title{Calculate USHE Element s_46 (Major College)}
\usage{
s_46(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (college_desc).}
}
\value{
Original data frame, with USHE data element s_46 appended.
}
\description{
Calculate USHE Element s_46 (Major College)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_46()

}
