% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_48}
\alias{s_48}
\title{Calculate USHE Element s_48 (Major College for Second Major)}
\usage{
s_48(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (secondary_major_college_id).}
}
\value{
Original data frame, with USHE data element s_48 appended.
}
\description{
Calculate USHE Element s_48 (Major College for Second Major)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_48()

}
