% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_50}
\alias{s_50}
\title{Calculate USHE Element s_50 (Degree Intent of Second Degree)}
\usage{
s_50(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (secondary_ipeds_award_level_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_50 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_50 (Degree Intent of Second Degree)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Degree Intent of Second Degree
\item FIELD NAME: S_DEG_INTENT2
\item FIELD FORMAT: Varchar, 2 Character
\item DEFINITION: The length/nature of the degree that the student is working toward.
S_DEG_INTENT2 with S_CURR_CIP2 represents the student’s major and ties to S_MAJOR2
}
}
\examples{
s_50()

}
