% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_xx}
\alias{s_xx}
\alias{s_alias}
\title{Calculate USHE Element s_xx (Name of Element)}
\usage{
s_xx(input_df = usheUtils::fake_student_df)

s_alias(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (list, of, required, data, fields).}
}
\value{
Original data frame, with USHE data element s_xx appended.
}
\description{
Calculate USHE Element s_xx (Name of Element)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_xx()

s_alias()
}
