% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_05}
\alias{sc_05}
\title{Calculate USHE Element sc_05 (Course Number)}
\usage{
sc_05(input_df = usheUtils::fake_student_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_number).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element sc_05 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element sc_05 (Course Number)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_05()

}
