% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_08}
\alias{sc_08}
\title{Calculate USHE Element sc_08 (Earned Credit Hours)}
\usage{
sc_08(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (earned_credits).}
}
\value{
Original data frame, with USHE data element sc_08 appended.
}
\description{
Calculate USHE Element sc_08 (Earned Credit Hours)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Credit Hours Earned
\item FIELD NAME: SC_EARNED_CR
\item FIELD FORMAT: Numeric (3,1)
\item DEFINITION: Credit hours earned by the student for the course. This field is for credit courses only.
}
}
\examples{
sc_08()

}
