% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_09}
\alias{sc_09}
\title{Calculate USHE Element sc_09 (Non-Credit Contact Hours)}
\usage{
sc_09(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (earned_credits, course_level_id).}
}
\value{
Original data frame, with USHE data element sc_09 appended.
}
\description{
Calculate USHE Element sc_09 (Non-Credit Contact Hours)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_09()

}
