% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_10}
\alias{sc_10}
\title{Calculate USHE Element sc_10 (Course Grade)}
\usage{
sc_10(input_df = usheUtils::fake_student_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (final_grade).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element sc_10 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element sc_10 (Course Grade)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_10()

}
