% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_12}
\alias{sc_12}
\title{Calculate USHE Element sc_12 (Student Type)}
\usage{
sc_12(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (latest_student_type_code, budget_code, is_concurrent_course, section_number).}
}
\value{
Original data frame, with USHE data element sc_12 appended.
}
\description{
Calculate USHE Element sc_12 (Student Type)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Student Type Code
\item FIELD NAME: SC_STUDENT_TYPE
\item FIELD FORMAT: Varchar, 2 Characters
\item DEFINITION: The specific code which would identify students according to special registration/tuition status.
Average Daily Membership (ADM) is a count of Utah public education students that is taken at different times of the year to satisfy local, state, and federal data collection  needs and also to ensure that school districts are adequately funded, according to student population.
The Concurrent Enrollment Master List is a list of college-level courses which have been approved by the Utah State Board of Education for Concurrent Enrollment.
}
}
\examples{
sc_12()

}
