% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_13}
\alias{sc_13}
\title{Calculate USHE Element sc_13 (Institutionally Assigned ID)}
\usage{
sc_13(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (student_id).}
}
\value{
Original data frame, with USHE data element sc_13 appended.
}
\description{
Calculate USHE Element sc_13 (Institutionally Assigned ID)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_13()

}
