% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_14}
\alias{sc_14}
\title{Calculate USHE Element sc_14 (Course Reference Number)}
\usage{
sc_14(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_reference_number).}
}
\value{
Original data frame, with USHE data element sc_14 appended.
}
\description{
Calculate USHE Element sc_14 (Course Reference Number)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_14()

}
