% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_15}
\alias{sc_15}
\title{Calculate USHE Element sc_15 (Student Credit Hour Type)}
\usage{
sc_15(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_level_id).}
}
\value{
Original data frame, with USHE data element sc_15 appended.
}
\description{
Calculate USHE Element sc_15 (Student Credit Hour Type)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_15()

}
