% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_11}
\alias{c_11}
\title{Calculate USHE Element c_11 (Budget Code)}
\usage{
c_11(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (budget_code).}
}
\value{
Original data frame, with USHE data element c_11 appended.
}
\description{
Calculate USHE Element c_11 (Budget Code)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_11()

}
