% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_14}
\alias{c_14}
\title{Calculate USHE Element c_14 (Credit Indicator)}
\usage{
c_14(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (course_level_id).}
}
\value{
Original data frame, with USHE data element c_14 appended.
}
\description{
Calculate USHE Element c_14 (Credit Indicator)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_14()

}
