% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_18}
\alias{c_18}
\title{Calculate USHE Element c_18 (Building Short Name)}
\usage{
c_18(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_building_id_1).}
}
\value{
Original data frame, with USHE data element c_18 appended.
}
\description{
Calculate USHE Element c_18 (Building Short Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_18()

}
