% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_40}
\alias{c_40}
\title{Calculate USHE Element c_40 (Course End Date)}
\usage{
c_40(input_df = usheUtils::fake_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (meet_end_date).}
}
\value{
Original data frame, with USHE data element c_40 appended.
}
\description{
Calculate USHE Element c_40 (Course End Date)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_40()

}
