% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\name{c_49}
\alias{c_49}
\title{Calculate USHE Element c_49 (Class Size)}
\usage{
c_49(input_df = usheUtils::fake_course_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (class_size).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element c_49 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element c_49 (Class Size)
}
\details{
\strong{USHE Documentation}
}
\examples{
c_49()

}
