% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{generate_student_course_submission_file}
\alias{generate_student_course_submission_file}
\title{Generate Students Courses Submission File}
\usage{
generate_student_course_submission_file(
  input_df = usheUtils::fake_student_course_df
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (subject_code,
course_number,
section_number,
attempted_credits,
earned_credits,
course_level_id,
final_grade,
latest_student_type_code,
budget_code,
is_concurrent_course,
student_id,
ssn,
course_reference_number).}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
}
\description{
Generate Students Courses Submission File
}
\examples{
generate_student_course_submission_file()

}
