% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{generate_student_submission_file}
\alias{generate_student_submission_file}
\title{Generate Student Submission File}
\usage{
generate_student_submission_file(
  input_df = usheUtils::fake_student_df,
  with_intermediates = FALSE
)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (season, academic_year_code, version_id, student_id,
ssn, ssid, previous_student_id,
last_name, first_name, middle_name, name_suffix,
previous_last_name, previous_first_name, previous_middle_name,
previous_name_suffix, local_address_zip_code, mailing_address_zip_code,
us_citizenship_code, first_admit_county_code, first_admit_state_code,
first_admit_country_iso_code, first_admit_state_code, birth_date, gender_code,
is_hispanic_latino_ethnicity, is_asian, is_black, is_american_indian_alaskan,
is_hawaiian_pacific_islander, is_white, is_international, is_other_race,
residency_code, primary_major_cip_code, student_type_code,
primary_level_class_id, primary_degree_id, level_id,
institutional_cumulative_credits_earned, institutional_cumulative_gpa,
transfer_cumulative_credits_earned, total_cumulative_clep_credits_earned,
total_cumulative_ap_credits_earned, full_time_part_time_code, version_date,
birth_date, first_admit_country_iso_code, high_school_code,
house_bill_75_waiver, secondary_major_cip_code, act_composite_score,
primary_major_cip_code, act_english_score, act_math_score, act_reading_score,
act_science_score, high_school_graduation_date, institutional_gpa,
is_pell_eligible, is_pell_awarded, is_bia, primary_major_college_id,
primary_major_desc, secondary_major_college_id, secondary_major_desc,
institutional_term_gpa, college_desc,
secondary_ipeds_award_level_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
A Data Frame, with all of the USHE elements required for upload submission.
This will also include intermediate values, used to calculate USHE data elements, if option is set.
}
\description{
Generate Student Submission File
}
\examples{
generate_student_submission_file()

}
