% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missionary.R
\name{m_02}
\alias{m_02}
\title{Calculate USHE Element m_02 (Student Name)}
\usage{
m_02(input_df = usheUtils::fake_mission_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (last_name, first_name, middle_name).}
}
\value{
Original data frame, with USHE data element m_02 appended.
}
\description{
Calculate USHE Element m_02 (Student Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
m_02()

}
