% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missionary.R
\name{m_04}
\alias{m_04}
\title{Calculate USHE Element m_04 ()}
\usage{
m_04(input_df = usheUtils::fake_mission_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (end_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element m_04 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element m_04 ()
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME:
\item FIELD NAME: m_end_date
\item FIELD FORMAT:
\item DEFINITION:
}
}
\examples{
m_04()

}
