% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_04}
\alias{pf_04}
\title{Calculate USHE Element pf_04 (Degree Level)}
\usage{
pf_04(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ipeds_award_level_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_04 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_04 (Degree Level)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_04()

}
