% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_05}
\alias{pf_05}
\title{Calculate USHE Element pf_05 (Degree Type)}
\usage{
pf_05(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (degree_id).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_05 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_05 (Degree Type)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_05()

}
