% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_08}
\alias{pf_08}
\title{Calculate USHE Element pf_08 (Minimum Required Credit Hours)}
\usage{
pf_08(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (required_credits).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_08 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_08 (Minimum Required Credit Hours)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_08()

}
