% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{pf_10}
\alias{pf_10}
\title{Calculate USHE Element pf_10 (Major Name)}
\usage{
pf_10(input_df = usheUtils::fake_program_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (major_desc).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element pf_10 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element pf_10 (Major Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
pf_10()

}
