% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_06}
\alias{s_06}
\title{Calculate USHE Element s_06 (Student Name)}
\usage{
s_06(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (last_name, first_name, middle_name, name_suffix).}
}
\value{
Original data frame, with USHE data element s_06 appended.
}
\description{
Calculate USHE Element s_06 (Student Name)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_06()

}
