% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_09}
\alias{s_09}
\title{Calculate USHE Element s_09 (Citizenship Code)}
\usage{
s_09(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (us_citizenship_code).}
}
\value{
Original data frame, with USHE data element s_09 appended.
}
\description{
Calculate USHE Element s_09 (Citizenship Code)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Citizenship Status
\item FIELD NAME: S_CITZ_CODE
\item FIELD FORMAT: Varchar, 1 Character
\item DEFINITION: Indicates student's current citizenship status.
}
}
\examples{
s_09()

}
