% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_12}
\alias{s_12}
\alias{m_03}
\title{Calculate USHE Element s_12 (Birth Date)}
\usage{
s_12(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)

m_03(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (birth_date).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_12 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_12 (Birth Date)
}
\details{
\strong{USHE Documentation}
\itemize{
\item ELEMENT NAME: Date of Birth
\item FIELD NAME: S_BIRTH_DT
\item FIELD FORMAT: Varchar, 8 Characters (YYYYMMDD),
\item DEFINITION: The calendar student date of birth, as designated by student.
}
}
\examples{
s_12()

m_03()
}
