% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_28}
\alias{s_28}
\title{Calculate USHE Element s_28 (High School Codes)}
\usage{
s_28(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (high_school_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_28 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_28 (High School Codes)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_28()

}
