% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_30}
\alias{s_30}
\title{Calculate USHE Element s_30 (CIP Code for Second Major)}
\usage{
s_30(input_df = usheUtils::fake_student_df, with_intermediates = FALSE)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (secondary_major_cip_code).}

\item{with_intermediates}{Boolean: Option to include intermediate calculated fields.}
}
\value{
Original data frame, with USHE data element s_30 appended. Will also return appended intermediate calculated fields, if option is set.
}
\description{
Calculate USHE Element s_30 (CIP Code for Second Major)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_30()

}
