% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_34}
\alias{s_34}
\title{Calculate USHE Element s_34, g_20 (SSID aka USOE Unique ID)}
\usage{
s_34(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (ssid).}
}
\value{
Original data frame, with USHE data element s_34 appended.
}
\description{
Calculate USHE Element s_34, g_20 (SSID aka USOE Unique ID)
}
\details{
\strong{USHE Documentation}
ELEMENT NAME: Unique USOE State Student Identifier
FIELD NAME: s_ssid, s_G_SSID
FIELD FORMAT: Varchar, 9 Characters
DEFINITION: The unique Utah State Student Identifier as assigned to each Utah public education student by the Utah State Board of Education.
This 9-digit number will appear on all high school transcripts beginning the 2006-2007 academic year.
}
\examples{
s_34()

}
