% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\name{s_40}
\alias{s_40}
\title{Calculate USHE Element s_40 (ACT Reading Sub-score)}
\usage{
s_40(input_df = usheUtils::fake_student_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (act_reading_score).}
}
\value{
Original data frame, with USHE data element s_40 appended.
}
\description{
Calculate USHE Element s_40 (ACT Reading Sub-score)
}
\details{
\strong{USHE Documentation}
}
\examples{
s_40()

}
