% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_07}
\alias{sc_07}
\title{Calculate USHE Element sc_07 (Attempted Credit Hours)}
\usage{
sc_07(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: (attempted_credits).}
}
\value{
Original data frame, with USHE data element sc_07 appended.
}
\description{
Calculate USHE Element sc_07 (Attempted Credit Hours)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_07()

}
