% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_courses.R
\name{sc_11}
\alias{sc_11}
\title{Calculate USHE Element sc_11 (Membership Hours)}
\usage{
sc_11(input_df = usheUtils::fake_student_course_df)
}
\arguments{
\item{input_df}{A Data Frame. Must contain the following data fields: ().}
}
\value{
Original data frame, with USHE data element sc_11 appended.
}
\description{
Calculate USHE Element sc_11 (Membership Hours)
}
\details{
\strong{USHE Documentation}
}
\examples{
sc_11()

}
